# Quic server batching