--- a/sdk/src/net.rs
+++ b/sdk/src/net.rs
@@ -1,0 +1,1 @@
+pub const DEFAULT_TPU_COALESCE_MS: u64 = 5;