@lijunwangs 2023-02-15
Can you post the bench-tpu results (before and after) here for reference?