@ryleung-solana 2023-02-15
> Can you post the bench-tpu results (before and after) here for reference?

Like just the screenshots? Sure

Before
![Screenshot from 2023-02-15 23-37-08](https://user-images.githubusercontent.com/91908731/219122214-6a20a3a8-2767-4da2-94b3-4097941ab40b.png)


After
![Screenshot from 2023-02-15 23-09-13](https://user-images.githubusercontent.com/91908731/219122240-c49b15ca-67a1-42c6-a949-f4aba28714cd.png)