@ryleung-solana 2023-02-15
> Something like these:
> 
> [#29774 (comment)](https://github.com/solana-labs/solana/pull/29774#issuecomment-1402369982)

How do you get those results :sweat_smile: I've just been relying on Graphana