@lijunwangs 2023-02-16
> > Something like these:
> > [#29774 (comment)](https://github.com/solana-labs/solana/pull/29774#issuecomment-1402369982)
> 
> How do you get those results 😅 I've just been relying on Graphana. Is it just grep output from the logs?

They are just bench-tps output in the end of run showing the aggregated performance. You can bound the runtime of  bench-tps like 1 min or 2 mins to see that. That gives us some single number to look at and compare.