@KirillLykov 2023-02-16
I wonder if we have some way to bench gossip in isolation?