@ryleung-solana 2023-02-17
> > > Something like these:
> > > [#29774 (comment)](https://github.com/solana-labs/solana/pull/29774#issuecomment-1402369982)
> > 
> > 
> > How do you get those results sweat_smile I've just been relying on Graphana. Is it just grep output from the logs?
> 
> They are just bench-tps output in the end of run showing the aggregated performance. You can bound the runtime of bench-tps like 1 min or 2 mins to see that. That gives us some single number to look at and compare.

Ok, got some results. Generally seems like no significant difference

After:
    Average max TPS: 102412.95, 0 nodes had 0 TPS
[2023-02-17T15:49:08.718469530Z INFO  solana_bench_tps::bench] 
    Highest TPS: 102412.95 sampling period 1s max transactions: 24705612 clients: 1 drop rate: 0.00
[2023-02-17T15:49:08.718474113Z INFO  solana_bench_tps::bench] 	Average TPS: 27260.873

    Average max TPS: 76727.48, 0 nodes had 0 TPS
[2023-02-17T15:49:15.209208740Z INFO  solana_bench_tps::bench] 
    Highest TPS: 76727.48 sampling period 1s max transactions: 23509636 clients: 1 drop rate: 0.00
[2023-02-17T15:49:15.209213578Z INFO  solana_bench_tps::bench] 	Average TPS: 25944.125

Before:
    Average max TPS: 104657.95, 0 nodes had 0 TPS
[2023-02-17T15:18:37.740772389Z INFO  solana_bench_tps::bench] 
    Highest TPS: 104657.95 sampling period 1s max transactions: 23950838 clients: 1 drop rate: 0.00
[2023-02-17T15:18:37.740790281Z INFO  solana_bench_tps::bench] 	Average TPS: 26538.877

    Average max TPS: 101332.98, 0 nodes had 0 TPS
[2023-02-17T15:18:32.461234128Z INFO  solana_bench_tps::bench] 
    Highest TPS: 101332.98 sampling period 1s max transactions: 22745441 clients: 1 drop rate: 0.00
[2023-02-17T15:18:32.461238175Z INFO  solana_bench_tps::bench] 	Average TPS: 25157.32