@KirillLykov 2023-02-17
I've setup a cluster with 3 nodes, 2 in toronto (1DC) and 1 in amsterdam (another DC), client in Amsterdam.

```
# before changes
 ./net.sh startclients  --client-type thin-client
mean_tps: 64688
 ./net.sh startclients  --client-type tpu-client
mean_tps: 65104
 ./net.sh startclients  --client-type rpc-client
mean_tps: 48015

# after changes
./net.sh startclients  --client-type thin-client
mean_tps: 45842
 ./net.sh startclients  --client-type tpu-client
mean_tps: 48855
 ./net.sh startclients  --client-type rpc-client
mean_tps: 45882
```

I've turned off CC to use one bench-tps.

Metrics looks quite weird unstable: on the left new code on the right old one
<img width="1184" alt="Screenshot 2023-02-17 at 20 17 42" src="https://user-images.githubusercontent.com/687962/219766757-fcfcf78f-0250-492e-a3ca-fe04ddc14bb7.png">