@KirillLykov 2023-02-19
Maybe not directly related to these changes, but important to me to understand how this code works:

In `handle_connection` we have a loop on stream exit:
```rust
while !stream_exit.load(Ordering::Relaxed) {
```

I don't understand if we ever exit it in case if we read all the chunks in the stream.

This flags is coming from `ConnectionTable` and stored inside of the `ConnectionEntry`.
It is set to `true` when this entry is dropped. 
I see it can be dropped when prune happens (new connection with higher stake).
But will this connection ever be stopped if all the packets from stream has been read?

Another thing is shouldn't `stream_exit` should be probably called `connection_exit`?