@ryleung-solana 2023-02-20
> In `handle_connection` we have a loop on stream exit:
> 
> ```rust
> while !stream_exit.load(Ordering::Relaxed) {
> ```
> 
> shouldn't `stream_exit` be called `connection_exit`?

Imho, it should really just be called exit or server_exit or something, but not a huge deal I guess...