@KirillLykov 2023-02-20
Results for UDP:
* b3778d5044 => 59k
* your branch => 44k