@ryleung-solana 2023-02-20
> Results for UDP:
> 
> * [b3778d5](https://github.com/solana-labs/solana/commit/b3778d5044dee5f441d7be7d5926841696111727) => 59k
> * your branch => 44k
> 
> One thing is that despite of using `--tpu-disable-quic` for client and `--tpu-disable-quic --tpu-enable-udp` for net.sh start, in metrics I see 150k of `quic-connections.packet_batches_sent`. Is it expected?

That's super weird... what args were used to run with UDP?