@KirillLykov 2023-02-20
> > Results for UDP:
> > 
> > * [b3778d5](https://github.com/solana-labs/solana/commit/b3778d5044dee5f441d7be7d5926841696111727) => 59k
> > * your branch => 44k
> > 
> > One thing is that despite of using `--tpu-disable-quic` for client and `--tpu-disable-quic --tpu-enable-udp` for net.sh start, in metrics I see 150k of `quic-connections.packet_batches_sent`. Is it expected?
> 
> That's super weird, this doesn't even touch the UDP path... what args were used to run with UDP?

```bash
#!/usr/bin/env bash

REPO_ROOT="./"
RESULT_FILE="result.txt"
INFLUX_HOST=https://internal-metrics.solana.com:8086
TESTNET_TAG="testnet-dev-klykov"
TEST_DURATION_SECONDS=600
# give it 5min to launch
TEST_STATS_SECONDS=300

function collect_mean_tps {
  declare q_mean_tps='
    SELECT ROUND(MEAN("median_sum")) as "mean_tps" FROM (
      SELECT MEDIAN(sum_count) AS "median_sum" FROM (
        SELECT SUM("count") AS "sum_count"
          FROM "'$TESTNET_TAG'"."autogen"."bank-process_transactions"
          WHERE time > now() - '"$TEST_STATS_SECONDS"'s AND count > 0
 GROUP BY time(1s), host_id)
      GROUP BY time(1s)
    )'

  curl -G "${INFLUX_HOST}/query?u=ro&p=topsecret" \
    --data-urlencode "db=${TESTNET_TAG}" \
    --data-urlencode "q=$q_mean_tps" |
    python3 "${REPO_ROOT}"../system-test/testnet-automation-json-parser.py >>"$RESULT_FILE"
}

# Amsterdam
CLIENT_IP="147.75.87.167"
# Amsterdam, Chicago, Toronto
NODE_IP="147.75.85.199"

function run_client {
  # thin
  echo " ./net.sh startclients  --client-type thin-client" >>"$RESULT_FILE"
  ./net.sh startclients  -c bench-tps=1=\"--tpu-disable-quic\"
  sleep $TEST_DURATION_SECONDS
  collect_mean_tps
  ./net.sh stopnode -i ${CLIENT_IP}
  
   # tpu
  #echo " ./net.sh startclients  --client-type tpu-client" >>"$RESULT_FILE"
  #./net.sh startclients  --client-type tpu-client
  #sleep $TEST_DURATION_SECONDS
  #collect_mean_tps
  #./net.sh stopnode -i ${CLIENT_IP}

  # rpc
  #echo " ./net.sh startclients  --client-type rpc-client" >>"$RESULT_FILE"
  #./net.sh startclients  --client-type rpc-client
  #sleep $TEST_DURATION_SECONDS
  #collect_mean_tps
  #./net.sh stopnode -i ${CLIENT_IP}
}

./init-metrics.sh -e kirill
./net.sh start --tpu-disable-quic --tpu-enable-udp
sleep 120

./net.sh stopnode -i ${CLIENT_IP}

# each run takes 10min, total ~50min
run_client
./net.sh stop
```