@ryleung-solana 2023-02-21
More comprehensive results for the newest rebased changes. Generally seems to be no significant perf impact, or, if anything, a slight advantage for the new changes. This is especially dramatic in the thin-client results, but TBH, I wouldn't rely too much on them, since even on the same branch, they can be a bit variable.

After:

TPU-Client

Average max TPS: 74737.05, 0 nodes had 0 TPS
[2023-02-21T17:50:42.690343755Z INFO  solana_bench_tps::bench] 
    Highest TPS: 74737.05 sampling period 1s max transactions: 23303971 clients: 1 drop rate: 0.00
[2023-02-21T17:50:42.690347987Z INFO  solana_bench_tps::bench] 	Average TPS: 25779.03

Average max TPS: 73994.77, 0 nodes had 0 TPS
[2023-02-21T17:50:36.990584832Z INFO  solana_bench_tps::bench] 
    Highest TPS: 73994.77 sampling period 1s max transactions: 23328541 clients: 1 drop rate: 0.00
[2023-02-21T17:50:36.990588964Z INFO  solana_bench_tps::bench] 	Average TPS: 25804.566

RPC-Client

Average max TPS: 28215.32, 0 nodes had 0 TPS
[2023-02-21T18:17:26.309633919Z INFO  solana_bench_tps::bench] 
    Highest TPS: 28215.32 sampling period 1s max transactions: 8953558 clients: 1 drop rate: 0.00
[2023-02-21T18:17:26.309637922Z INFO  solana_bench_tps::bench] 	Average TPS: 9842.38

Average max TPS: 26471.89, 0 nodes had 0 TPS
[2023-02-21T18:17:26.261087741Z INFO  solana_metrics::metrics] datapoint: bench-tps-lamport_balance balance=0i
[2023-02-21T18:17:26.261090292Z INFO  solana_bench_tps::bench] 
    Highest TPS: 26471.89 sampling period 1s max transactions: 8986439 clients: 1 drop rate: 0.00
[2023-02-21T18:17:26.261097447Z INFO  solana_bench_tps::bench] 	Average TPS: 9813.087

Thin-Client

Average max TPS: 111751.38, 0 nodes had 0 TPS
[2023-02-21T18:46:57.541682417Z INFO  solana_bench_tps::bench] 
    Highest TPS: 111751.38 sampling period 1s max transactions: 23471727 clients: 1 drop rate: 0.00
[2023-02-21T18:46:57.541686846Z INFO  solana_bench_tps::bench] 	Average TPS: 25961.227

Average max TPS: 116404.23, 0 nodes had 0 TPS
[2023-02-21T18:46:52.436226116Z INFO  solana_metrics::metrics] datapoint: bench-tps-lamport_balance balance=0i
[2023-02-21T18:46:52.436228576Z INFO  solana_bench_tps::bench] 
    Highest TPS: 116404.23 sampling period 1s max transactions: 22539406 clients: 1 drop rate: 0.00
[2023-02-21T18:46:52.436269317Z INFO  solana_bench_tps::bench] 	Average TPS: 24955.746

Before:

TPU-Client

Average max TPS: 71624.62, 0 nodes had 0 TPS
[2023-02-21T19:19:40.689737572Z INFO  solana_bench_tps::bench] 
    Highest TPS: 71624.62 sampling period 1s max transactions: 22359885 clients: 1 drop rate: 0.00
[2023-02-21T19:19:40.689736610Z INFO  solana_metrics::metrics] datapoint: bench-tps-lamport_balance balance=0i
[2023-02-21T19:19:40.689742046Z INFO  solana_bench_tps::bench] 	Average TPS: 24722.113

Average max TPS: 71643.05, 0 nodes had 0 TPS
[2023-02-21T19:19:35.328187181Z INFO  solana_bench_tps::bench] 
    Highest TPS: 71643.05 sampling period 1s max transactions: 22407744 clients: 1 drop rate: 0.00
[2023-02-21T19:19:35.328191849Z INFO  solana_bench_tps::bench] 	Average TPS: 24785.635

RPC-Client

Average max TPS: 25954.35, 0 nodes had 0 TPS
[2023-02-21T19:46:16.108313526Z INFO  solana_bench_tps::bench] 
    Highest TPS: 25954.35 sampling period 1s max transactions: 9436593 clients: 1 drop rate: 0.00
[2023-02-21T19:46:16.108317803Z INFO  solana_bench_tps::bench] 	Average TPS: 10373.474

Average max TPS: 26712.71, 0 nodes had 0 TPS
[2023-02-21T19:46:14.608987257Z INFO  solana_bench_tps::bench] 
    Highest TPS: 26712.71 sampling period 1s max transactions: 9446406 clients: 1 drop rate: 0.00
[2023-02-21T19:46:14.608991523Z INFO  solana_bench_tps::bench] 	Average TPS: 10338.99

Thin-Client

Average max TPS: 89094.77, 0 nodes had 0 TPS
[2023-02-21T20:13:17.262552955Z INFO  solana_bench_tps::bench] 
    Highest TPS: 89094.77 sampling period 1s max transactions: 18878731 clients: 1 drop rate: 0.00
[2023-02-21T20:13:17.262556842Z INFO  solana_bench_tps::bench] 	Average TPS: 20906.68

Average max TPS: 89084.66, 0 nodes had 0 TPS
[2023-02-21T20:13:11.896986416Z INFO  solana_metrics::metrics] datapoint: bench-tps-lamport_balance balance=0i
[2023-02-21T20:13:11.897018730Z INFO  solana_bench_tps::bench] 
    Highest TPS: 89084.66 sampling period 1s max transactions: 18629590 clients: 1 drop rate: 0.00
[2023-02-21T20:13:11.897023812Z INFO  solana_bench_tps::bench] 	Average TPS: 20618.26